Significantly improved performance by removing 1 symlink check
Emacs was constantly consuming 100% CPU whenever opening a new file
with many problems. The reason I found, by the use of profiler-report,
was in eclim-problems-highlight where an extra call to file-truename
was made to resolve any symlinks in the path to the Java at hand.

This change was introduced in b0368df2 to fix the handling of
symlinks, but it makes Emacs/eclim unusuable for me on large projects
where Emacs spends ~8 seconds to switch to an already opened Java
file.

With this wee change, Emacs opens & switches to Java-buffers
instantaneously and CPU normally doesn't go above 5% for the same
opening of a Java file and decoration of errors in that buffer.

In case anyone is interested: My system: GNU Emacs 24.4.1, Debian
Jessie, Intel i7 CPU, 8 GB