Remove redundant prefix from candidates
This is to fix the issue where if you complete 'java.uti' it will end up
inserting 'java.java.util'. This is because the completion prefix in
this example is 'uti', but the candidate returned by eclim will be the
full package name 'java.util'. Company will insert the completion from
the prefix, which leads to the incorrect behavior. This solves this
issue by looking before the prefix at what is already present in the
buffer, and removing and redundancy from candidates if present.

Closes #213