company-eclim: make ignore-case=nil filter candidates to match prefix.
Eclipse changed the relevant option to apply to camelCase only, and when
completing a variable we get a *huge* selection of all classes that begin
with that prefix. Try completing a variable that starts with "re" for example.
Eclimd's sorting is not particularly good, either -- in Eclipse at least we
get the lowercase matches first.

Company-mode's view is that the backend is responsible for ensuring the
prefix matches properly.