Added classpath features that eclim provides
User now has the ability to view source directories and eclipse
variables from within a specific buffer. From that buffer they are also
able to add/delete source directories from the .classpath file and
add/delete variables from eclipse as well. User is able to add a new
variable entry to the .classpath file as well through the use of a
independent function.