Problem highlights: add variable that causes highlights to be suppressed.
This allows the user to easily set up commands that disable highlights
(e.g.  temporarily, while writing a bunch of new code), globally or
buffer-local (more common in my usage).

Unlike a problem filter, the problem counts remain and the problems are
navigable.

Rename eclim--problems-clear-highlights to indicate it's not private.