Completion: fix up corner cases in yas template computation.
One was that nested parameters created a nested template, which did not
make sense (filling template args before erasing the template altogether);
it did not seem intended. The other was an empty arg list.

Instead, use replace-regexp-in-string which takes care of some ugly
details, and count levels to make sure we replace only on first.

Added a test. The previous code fails the first two assertions.