Open files in archives. Fixes http://github.com/senny/emacs-eclim/issues/250
The fix was twofold:
1) Fix regexp to allow jar:file:///Users...
2) Remove explicit filtering of results against archives. There was code to
handle them, but it had been explicitly disabled presumably because of the
problem above, which made it behave rather badly.

This works beautifully when source jars if present, including all of the
SDK. It also works with class files and autodisass-java-bytecode , though
personally I don't find that very useful; better than bytecode, I suppose.