Completion: allow the user to customize insertion.
I think yas is problematic for completion because:
* Completion begins by spitting out a potentially long method signature
* Aborting from yas can't resume, so you're left with above
* The code can be in a non-sensical (for Eclipse) state for quite
a long time. This can prevent further coding aids.
* Yas is intrusive and wants to be a minor mode everywhere.

The default insertion is hardly better. So one should at least have the
option to do something different. I put in a simple (and sample)
alternative that simply erases arguments, which is already an improvement,
but I have something better for the next commit.